update: menu functions; add: pay in chat, chat activity checking

+++ app/handlers/private/my_posts.py
@@ -11,35 +11,42 @@ from app.database.services.repos import PostRepo, DealRepo
 from app.handlers.private.start import start_cmd
 from app.keyboards import Buttons
 from app.keyboards.inline.post import construct_posts_list_kb, post_cb, moderate_post_kb, participate_kb, delete_post_kb
-from app.keyboards.reply.menu import basic_kb
 from app.misc.times import localize, now
 
 
 async def my_posts_cmd(msg: Message, post_db: PostRepo):
-    posts = await post_db.get_posts_user(msg.from_user.id, DealStatusEnum.ACTIVE)
+    posts = await post_db.get_posts_user(msg.from_user.id)
+    posts.sort(key=lambda post: post.created_at)
+    posts = [post for post in posts if post.status in (DealStatusEnum.ACTIVE, DealStatusEnum.BUSY,
+                                                       DealStatusEnum.MODERATE)]
     if not posts:
         await msg.answer('Спочатку опублікуйте хоча б один пост 🙂')
         return
-    posts.sort(key=lambda post: post.created_at)
     text = (
         f'📑 Ваші пости, опубліковані на каналі:\n\n'
         f'{construct_posts_list(posts)}\n'
     )
     await msg.answer(text, reply_markup=construct_posts_list_kb(posts))
-    text = (
-        f'Ви можете видалити або оновити пост в каналі, якщо він має статус '
-        f'"{PostStatusText.ACTIVE}".'
-    )
-    await msg.answer(text, reply_markup=basic_kb([Buttons.menu.back]))
 
 
 async def edit_post_cmd(call: CallbackQuery, callback_data: dict, post_db: PostRepo):
     post_id = int(callback_data['post_id'])
     post = await post_db.get_post(post_id)
-    await call.message.edit_text(
-        text=hide_link(post.post_url) + post.construct_post_text(use_bot_link=False),
-        reply_markup=moderate_post_kb(post)
+    if post.status != DealStatusEnum.ACTIVE:
+
+        if post.status == DealStatusEnum.MODERATE:
+            status = 'він ще не схвалений модератором'
+        elif post.status == DealStatusEnum.BUSY:
+            status = 'він вже виконується у чаті'
+        else:
+            status = 'він був відхилений модератором'
+        await call.answer(f'Ви не можете редагувати цей пост, оскільки {status}', show_alert=True)
+        return
+    text = (
+        f'{post.construct_post_text(use_bot_link=False)}\n\n'
+        f'{post.construct_html_link("Перейти до поста в каналі")}'
     )
+    await call.message.edit_text(text=text, reply_markup=moderate_post_kb(post))
 
 
 async def delete_post_cmd(call: CallbackQuery, callback_data: dict, post_db: PostRepo):
@@ -55,24 +62,25 @@ async def delete_post_cmd(call: CallbackQuery, callback_data: dict, post_db: Pos
 
 
 async def confirm_delete_post_cmd(call: CallbackQuery, callback_data: dict, post_db: PostRepo, deal_db: DealRepo,
-                                  config: Config, state: FSMContext):
+                                  config: Config):
     post_id = int(callback_data['post_id'])
     post = await post_db.get_post(post_id)
-    await call.bot.delete_message(
-        chat_id=config.misc.post_channel_chat_id, message_id=post.message_id
-    )
-    await call.bot.delete_message(
-        chat_id=config.misc.reserv_channel_id, message_id=post.reserv_message_id
-    )
+    if post.message_id:
+        await call.bot.delete_message(
+            chat_id=config.misc.post_channel_chat_id, message_id=post.message_id
+        )
+    if post.reserv_message_id:
+        await call.bot.delete_message(
+            chat_id=config.misc.reserv_channel_id, message_id=post.reserv_message_id
+        )
+    if post.admin_message_id:
+        await call.bot.delete_message(
+            chat_id=config.misc.admin_channel_id, message_id=post.admin_message_id
+        )
     await deal_db.delete_deal(post.deal_id)
     await post_db.delete_post(post_id)
     await call.answer('Пост було видалено', show_alert=True)
-    posts = await post_db.get_posts_user(call.from_user.id, DealStatusEnum.ACTIVE)
-    if posts:
-        await back_posts_list_cmd(call, post_db)
-    else:
-        await call.message.delete()
-        await start_cmd(call.message, state)
+    await back_posts_list_cmd(call, post_db)
 
 
 async def update_post_cmd(call: CallbackQuery, callback_data: dict, post_db: PostRepo, config: Config):
@@ -92,8 +100,8 @@ async def update_post_cmd(call: CallbackQuery, callback_data: dict, post_db: Pos
         await edit_post_cmd(call, callback_data, post_db)
     else:
         text = (
-            f'Ви можете оновлювати пост в каналі не раніше ніж 15 хв після його публікації. '
-            f'Зачекайте {(15 * 60 - seconds_after_public) // 60} хв...'
+            f'Ви можете оновлювати пост в каналі не раніше ніж 15 хв після його публікації '
+            f'(залишилось {(15 * 60 - seconds_after_public) // 60} хв...)'
         )
         await call.answer(text, show_alert=True)
 
@@ -104,17 +112,17 @@ async def close_posts_cmd(call: CallbackQuery, state: FSMContext):
 
 
 async def back_posts_list_cmd(call: CallbackQuery, post_db: PostRepo):
-    posts = await post_db.get_posts_user(call.from_user.id, DealStatusEnum.ACTIVE)
+    posts = await post_db.get_posts_user(call.from_user.id)
+    posts.sort(key=lambda post: post.created_at)
+    posts = [post for post in posts if post.status in (DealStatusEnum.ACTIVE, DealStatusEnum.BUSY,
+                                                       DealStatusEnum.MODERATE)]
     if not posts:
+        await call.message.delete()
         await call.message.answer('Спочатку опублікуйте хоча б один пост 🙂')
         return
-    posts = [post for post in posts if post.message_id]
-    posts.sort(key=lambda post: post.created_at)
     text = (
         f'📑 Ваші пости, опубліковані на каналі:\n\n'
         f'{construct_posts_list(posts)}\n'
-        f'Ви можете видалити або оновити пост в каналі, якщо він має статус '
-        f'"{PostStatusText.ACTIVE}".'
     )
     await call.message.edit_text(text, reply_markup=construct_posts_list_kb(posts))
